Move ConvertResourceToARMResource() onto AzureDeploymentReconciler
